+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- Normalized event signature verification after failed raw content attempt
 
 ## [2.9.0] - 2021-03-22
 ### Added